+++ src/Helpers/Router.php
@@ -132,8 +132,8 @@ class Router
      */
     private function showConfiguration()
     {
-        \include_file('core', 'authentification', 'php');
-        \include_file('plugin_info', 'configuration', 'configuration', Utils::init('plugin'), true);
+        FileSystemHelper::includeFile('core', 'authentification', 'php');
+        FileSystemHelper::includeFile('plugin_info', 'configuration', 'configuration', Utils::init('plugin'), true);
     }
 
     /**
@@ -155,7 +155,7 @@ class Router
             $filename = $_GET['p'];
             $plugin = isset($_GET['m']) ? $_GET['m'] : $plugin;
         }
-        \include_file('mobile', $filename, $type, $plugin, true);
+        FileSystemHelper::includeFile('mobile', $filename, $type, $plugin, true);
     }
 
     /**
