+++ src/Managers/DevicesManager.php
@@ -35,6 +35,8 @@
 namespace NextDom\Managers;
 
 use NextDom\Exceptions\CoreException;
+use NextDom\Helpers\FileSystemHelper;
+use NextDom\Helpers\Utils;
 
 require_once NEXTDOM_ROOT.'/core/class/cache.class.php';
 
@@ -52,7 +54,7 @@ class DevicesManager {
         $cache = CacheManager::byKey('nextdom::usbMapping');
         if (!is_json($cache->getValue()) || $name == '') {
             $usbMapping = array();
-            foreach (\ls('/dev/', 'ttyUSB*') as $usb) {
+            foreach (FileSystemHelper::ls('/dev/', 'ttyUSB*') as $usb) {
                 $vendor = '';
                 $model = '';
                 $devsList = shell_exec('/sbin/udevadm info --name=/dev/' . $usb . ' --query=all');
@@ -98,7 +100,7 @@ class DevicesManager {
                 if (file_exists('/dev/ttyS1')) {
                     $usbMapping['Odroid C2'] = '/dev/ttyS1';
                 }
-                foreach (ls('/dev/', 'ttyACM*') as $value) {
+                foreach (FileSystemHelper::ls('/dev/', 'ttyACM*') as $value) {
                     $usbMapping['/dev/' . $value] = '/dev/' . $value;
                 }
             }
